Extract ZIP handler classes to separate files
